backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.22.0
httpx==0.24.1
pydantic==2.4.0
python-dotenv==1.0.0

# Database & migrations
sqlalchemy[asyncio]>=2.0,<3.0
asyncpg>=0.29,<1.0
alembic>=1.13,<2.0